adding redux saga, complete register, login to server

+++ utils/removeImage.js
@@ -2,6 +2,9 @@ const path = require("path");
 const fs = require("fs");
 const removeImage = (filename) => {
   return new Promise((resolve, reject) => {
+    if (filename === "avatar-default.png") {
+      return resolve(true);
+    }
     fs.unlink(
       path.join(path.dirname(require.main.filename), "images", filename),
       (error) => {
